Voice-Assistant-main/requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
pygame==2.6.1
requests==2.31.0
gTTS==2.5.1
openai==1.30.1
python-dotenv==1.0.1
pocketsphinx==5.0.4
